requirements.txt
Django>=3.2
Pillow
torch
torchvision
